requirements.txt
pandas
scikit-learn
numpy
pyyaml
dvc
joblib
pytest
flake8